requirements.txt
torch
torchvision
numpy
scipy
Pillow
nltk
spacy
keras
tensorflow
pycocotools
symspellpy
tqdm
inflect
